refactor nodeData & improve loading robustness

+++ src/main.ts
@@ -3,20 +3,19 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import 'bootstrap/dist/js/bootstrap.bundle.min.js';
 import "./styles/custom.scss";
 import "./styles/custom.css";
-import { Player } from "./creatures/Player";
 import { loadPlayer } from "./save";
 import { renderStartPage } from "./pages/startPages";
 import { renderMainMenu } from "./pages/mainMenu";
 import { renderMapPage } from "./pages/mapPage";
 import { CreatureType } from "./creatures/creatureConfigs";
+import { initializeAllMapData } from "./maps/Region";
+
+initializeAllMapData();
 
 // 初始加载时显示主菜单
 document.addEventListener("DOMContentLoaded", () => {
-  let player: Player;
-  try {
-    player = loadPlayer();
-  } catch (error) {
-    console.log(error);
+  const player = loadPlayer();
+  if (!player) {
     localStorage.clear();
     renderStartPage();
     return;
